Rewrite TextArea
